+++ CHANGELOG.md
@@ -72,6 +72,7 @@ Changes that are not related to specific components
 #### Fixed
 
 - Fixed pages with broken code examples
+- Fixed anchor links in documentation site
 
 ### Figma